requirements.txt
streamlit>=1.29.0
yfinance>=0.2.18
pandas>=2.1.0
numpy>=1.24.0
plotly>=5.17.0
# ta-lib==0.4.28  # Will add manual indicators
python-telegram-bot>=20.7
twilio>=8.10.0
gspread>=5.12.0
oauth2client>=4.1.3
speechrecognition>=3.10.0
# pyaudio==0.2.11  # Optional for advanced voice
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytz>=2023.3
websocket-client>=1.6.4
